requirements.txt
fastapi
uvicorn
pydantic
ollama
sentence-transformers
faiss-cpu
numpy